% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_gain.R
\name{information_gain}
\alias{information_gain}
\title{Calculate information gain}
\usage{
information_gain(splitted_labels)
}
\arguments{
\item{splitted_labels}{A list. Each element of it is a factor of labels.}
}
\description{
Based on changed entropy of the data, calculated information gain.
}
\details{
It is assumed, that y are splitted labels of a greater dataset.
When combined together, they should form a vector equivalent to labels in original dataset (order may differ).
}
