% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_value.R
\name{IV}
\alias{IV}
\title{Calculate information value}
\usage{
IV(splitted_labels)
}
\arguments{
\item{splitted_labels}{A list of binary factors}
}
\description{
Calculate information value
}
\section{Details}{

It is assumed, that y are splitted labels of a greater dataset.
When combined together, they should form a vector equivalent to labels in original dataset (order may differ).
}

