% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE.R
\name{SAFE}
\alias{SAFE}
\title{Scalable Automatic Feature Engineering}
\usage{
SAFE(
  X_train,
  y_train,
  X_valid,
  y_valid,
  operators = list(NULL, list(`+`, `-`, `*`, `/`)),
  n_iter,
  nrounds
)
}
\arguments{
\item{X_train}{matrix}

\item{y_train}{factor}

\item{X_valid}{matrix}

\item{y_valid}{factor}

\item{operators}{A \code{list} of lists of functions. Ith list of funcitons contains functions accepting i vectors of equal length and returning 1 vector of the same length}

\item{n_iter}{Integer}

\item{nrounds}{Integer for \code{\link[xgboost]{xgboost}}}
}
\description{
More description here
}
